feat(odm): NIXL-based consistency and seeding, 64-bit IOVA, remove ODM_ADDR
Replace host ioctl paths in nixlbench with NIXL transfers:
- Consistency (WRITE): NIXL READ device->host instead of host READ ioctl
- Seeding (READ): NIXL WRITE host->device instead of host WRITE ioctl
- Remove resolveDeviceIovas, queryMem, seedViaHostWrite from nixlbench
- Remove all odm_ioctl.h usage from benchmark code

Widen GET_IOVA/FREE_IOVA ioctl to 64-bit:
- Plugin: uint32_t target_iova_size -> uint64_t (removes 4GB limit)
- Remove UINT32_MAX size check in registerMem

Remove hardcoded ODM_ADDR:
- Remove ODM_ADDR env var from nixlbench, docs, configs
- Plugin auto-allocates IOVA via GET_IOVA when addr=0

Tested: 8GB buffer, 4KB-32MB blocks, batch=64, consistency pass
  WRITE peak: 25.15 GB/s  READ peak: 43.38 GB/s

+++ src/plugins/marvell_odm/odm_ioctl.h
@@ -80,7 +80,7 @@ struct mrvl_dma_xfer_commands {
 
 struct mrvl_dma_iova_commands {
     uint64_t target_iova_addr;
-    uint32_t target_iova_size;
+    uint64_t target_iova_size;
 };
 
 #define MRVL_CXL_DMA_READ_COMMAND _IOWR(ODM_IOCTL_MAGIC, 3, struct mrvl_dma_xfer_commands)
